fixed earlier errors with invalid inputs
Lowercase letters is now automatically checked so users will less likely make anerror of not checking a box. If no boxes are checked, it will notify the user and the function will only run if they have at least checked one box.

Tested for invalid input of no boxes checked and it works as it should. Program does not break