added a second function
I needed to have a function that returned a value for ncea so here you go.

now makepassword function is isolated and doesn't use variables from outside the function

tested for expected inputs of lengths 6, 8, 10, 12
tested for boundary inputs of lengths 6 & 20

tested for expected inputs of checkboxes lowercase, uppercase, numbers, and special characters.
tested invalid input of no checkboxes.

works as it should